DESENG-483: Fixing review comments
Moving common jest mock and spy to another file
